+++ src/main/java/com/hunteryavitz/blockchainapi/utils/TransactionUtils.java
@@ -1,4 +1,8 @@
 package com.hunteryavitz.blockchainapi.utils;
 
+/**
+ * The TransactionUtils class provides utility methods for transactions.
+ */
 public class TransactionUtils {
+
 }