+++ CHANGELOG.md
@@ -3,10 +3,13 @@ Unreleased
 
 *v0.0.1*
 
+**8/30/23**
+
+- [ADD Transaction to Block](#36)
+
 **8/27/23**
 
 - [Fix Failing Unit Test for Verify Blockchain](#34)g
-- [ADD Transaction to Block](#36)
 
 **8/24/23**